Add README refs for Lambda events from SES
While I've successfully used SES to send emails with the
`List-Unsubscribe` header and `{{amazonSESUnsubscribeUrl}}`, the
resulting URLs are uug-lee. Plus I don't like that the hostnames, while
legit Amazon hosts, are generic and sketchy, as is the SES subscription
management page.

As it turns out, SES can trigger a Lambda function on incoming email
events. So what I'll eventually do is compose raw emails that set
`List-Unsubscribe` and `{{ UnsubscribeUrl }}` to my own custom values.
For the `mailto:` version of `List-Unsubscribe` in particular, I'll set
up a second Lambda function to respond to SES events. I'll be able to
share the same underlying unsubscribe logic between the two Lambda
functions. I'll also use CloudFormation to set up the SES rule set to
trigger the Lambda.

I'll switch the subscriber list to a more persistent, not SAM- or
CloudFormation-controlled DynamoDB table. This is because there's no
other way to attach metadata like verification UUIDs to contact entities
in a SES contact list.

It's a little extra work, but not too much, and I think it'll be much
better user experience.