Move main.go to lambda/main.go, update deploy.sh
Moving main.go to lambda/main.go was inspired by:
- https://ieftimov.com/posts/golang-package-multiple-binaries/

This makes things a little more organized, especially if/when I create a
binary tool for sending messages with SES. The reason being, this tool
could automatically apply correctly formatted https: and mailto:
unsubscribe URIs generated by the subscription management system to
templates and headers.

My personal motivation for creating such a tool is that I have a
separate Ruby script to send email from my latest blog post. I can make
that script substitute the URIs in the correct format, but the idea of
keeping them in sync bothers me. I could keep the part of the Ruby
script that parses and formats the blog post in place. It would then
pass the text and html messages to the Go driver from this project
instead of to the AWS SDK.