Rename to EListMan, extensively update README
My original idea was "ListServerless," but apparently "LISTSERV" is
actually a registered trademark, so I decided not to risk it. "EListMan"
was a registered trademark once, but marked abandoned in 2005, so I took
it. I also like the idea of adding an `elistman` command suite in the
near future.

The README also now contains extensive information on prerequisites and
the URI schema. Updated the "Algorithms" section, and added the
/unsubscribe algorithm.