Define SES receipt rule for mailto:unsubscribe
We have to define EListManReceiptRuleSetPermission manually, since
there's no SES EventSource type for AWS::Serverless::Function.

This took a bit of trial and error, and necessitated adding a DependsOn
property to EListManUnsubscribeReceiptRule to make sure the permissions
were in place first. Otherwise the deployment would fail because the
receipt rule couldn't execute the EListMan Lambda function.

I also replaced the SenderEmailAddress parameter and the expression to
parse out the domain name with the EmailDomainName parameter. This made
it easier to use the email domain name in multiple places, including in
EListManUnsubscribeReceiptRule. For the time being, this also means that
the sender username is fixed as "no-reply" and the unsubscribe recipient
user name is fixed as "unsubscribe". If future needs demand flexibility,
I'll add it then, but these seem pretty reasonable, universal, and
durable defaults.

And while I was at it, I marked all the Makefile targets as .PHONY, to
make all the rules consistent. The actual build depenedency work is
handled by the go tool itself.