Test and fix Event.UnmarshalJSON, rm getEndpoint
In the process of testing Event.UnmarshalJSON(), I deployed an earlier
version of this change that returned 500s from bin/smoke-test.sh. The
fix was not to expect specific characters after the '"rawPath":' or
'"ses":' tokens.

However, while looking at the CloudWatch logs for the output from the
500, I was surprised by the "routeKey":

```json
{
    "version": "2.0",
    "routeKey": "POST /subscribe",
    "rawPath": "/subscribe",
    "rawQueryString": "",
    "headers": {
        "accept": "*/*",
        "content-length": "20",
        "content-type": "application/x-www-form-urlencoded",
        "host": "api.mike-bland.com",
        "user-agent": "curl/7.86.0",
        "x-amzn-trace-id": "<...>",
        "x-forwarded-for": "<...>",
        "x-forwarded-port": "443",
        "x-forwarded-proto": "https"
    },
    "requestContext": {
        "accountId": "<...>",
        "apiId": "<...>",
        "domainName": "api.mike-bland.com",
        "domainPrefix": "api",
        "http": {
            "method": "POST",
            "path": "/subscribe",
            "protocol": "HTTP/1.1",
            "sourceIp": "<...>",
            "userAgent": "curl/7.86.0"
        },
        "requestId": "<...>",
        "routeKey": "POST /subscribe",
        "stage": "$default",
        "time": "23/Mar/2023:18:18:59 +0000",
        "timeEpoch": 1679595539338
    },
    "body": "<...>",
    "isBase64Encoded": true
}
```

So by the time the Lambda receives the request, the API Gateway RouteKey
has been stripped from the RawPath. That meant that the getEndpoint
helper in handler.go was unnecessary, so I've removed it in this commit.