Find commonHeaders in SES event, no unmarshal err
Searching for '"commonHeaders":' to detect an SES event seems even more
rigorous.

After thinking about my idea from the previous commit, I realized we
shouldn't error out during event unmarshaling. I sketched out a full set
of parsing rules, but realized it's better to parse in the specific
LambdaHandler functions. If we error out during unmarshaling, the Lambda
returns a 500. If we detect errors during actual event handling, we can
return errors to the user on invalid requests as intended. (Or, in the
case of mailto events, log the fact that we do nothing.)