Extract parser functions, use mail.ParseAddress
This makes any given function a bit easier to follow and to test, and
mail.ParseAddress enables requests to fail fast, before dispatching to
operations. Also added LambdaHandler.prepareParseErrorResponse to
treat email parse errors differently for the /subscribe operation, since
it may be due to a user typo. In all other cases, the assumption is that
it's a bad machine generated request.