Create options.go, define all parms/env vars
All of the application's parameters are defined and required. The
GetOptions function parses them from environment variables, and returns
and error if any are undefined.

By the way, 'sam local' doesn't properly resolve all of the '!Ref
ParameterName' values used to define the env vars. 'sam deploy' does,
however. Good to know and to expect this surprising bit of behavior.
I've not yet found a definitive solution to make them resolve locally.