Remove '-coverpkg ./...' from 'make test'
I picked this flag up from another project that covered code in one
package with tests from another. Arguably the right thing would've been
to move the code or tests closer. However, this enabled me to ensure my
change was properly covered without suggesting a larger refactoring.

In this project, however, that option would hide problems that I want to
solve. Plus, it causes the tests to report 0% coverage for the last
package for some reason.