CI: Check go.mod is tidy and install dependencies
Given the output from the first version of ci.yml, it seems `go get .`
only installs the dependencies from tools.go in the top-level directory.
With this change, it will try installing all the packages in the
project, after checking that go.mod is up to date via 'go tidy'.