Run 'go get' before 'go mod tidy' in ci.yml
If the build runs on an instance that doesn't yet have a cache, 'go mod
tidy -x' will download modules and print whatever it downloads. So we
need to try to 'go get' the modules already in 'go.mod' and 'go.sum'
before making sure 'go mod tidy' is a nop.