Finish testing and refactoring parseApiEvent
The implementation grew a bit, but I'm happy with the result. I think
the responsibilities are better divided, the data organized, and the
logic clearer.

I was inspired by the example from the "Customizing error tests with Is
and As methods" section of https://go.dev/blog/go1.13-errors to try
ParseError.Is(). This feels better than defining many custom error
types, or checking for a particular error string in
Handler.prepareParseErrorResponse.

Plus, I got to try my hand at a generic function in Go for the first
time!