Add OperationResult for SubscriptionAgent returns
It's clear that just returning a bool from each of the SubscriptionAgent
methods isn't enough information, and that anticipated use cases aren't
exactly errors. Infrastructure errors, on the other hand, should be
returned as proper error types.

While it's possible to restrict each of the SubscriptionAgent methods to
return separate types, I'm not sure there's sufficient benefit to
justify the effort.

This change sets the stage for the Handler to create a map from
OperationResult to RedirectUrls, making the redirect dispatching logic
very straightforward.