Add handleApiEvent, logApiResponse, op error types
The HTTP implemention now guarantees that all errors are reported via
the HTTP response, not by returning an error to the Lambda framework.
All HTTP requests are now logged by the Lambda function, in addition to
any logging by the API Gateway.

In the process, I learned why it's always good practice to return
"error" instead of a custom error type:

- "Go FAQ: Why is my nil error value not equal to nil?"
  https://go.dev/doc/faq#nil_error