Set SES event verdicts, DmarcPolicy to uppercase
Notice that according to:
- https://docs.aws.amazon.com/ses/latest/dg/receiving-email-action-lambda-example-functions.html

all of the verdicts and the DMARCPolicy should be all uppercase.

However, according to:
- https://docs.aws.amazon.com/ses/latest/dg/receiving-email-notifications-contents.html

The DMARCPolicy should be all lowercase. As a defensive measure, we
explicitly uppercase them all.

simpleEmailService() from handler_test.go sets all verdicts and
DMARCPolicy to lowercase to make sure that TestNewMailtoEvent validates
that newMailtoHandler() uppercases them all.