Add email.Bouncer interface, stub, and test double
This is in anticipation of bouncing messages due to DMARC failures, per:

https://docs.aws.amazon.com/ses/latest/dg/receiving-email-action-lambda-example-functions.html

I thought of adding a new passthrough method to ops.SubscriptionAgent
instead, but decided it might be better to define a Bouncer
interface. Since bouncing an email is very specific to the
mailtoHandler, it seemed best to keep that behavior separate from the
SubscriptionAgent interface. In contrast, the SubscriptionAgent is
relevate to the mailtoHandler (for Unsubscribe), to the apiHandler (all
operations), and to the upcoming command line interface.

However, SesMailer implements both the Mailer and Bouncer interfaces,
since it serves as the internal API to the external AWS SES API.