Add --no-confirm-changeset to 'sam deploy'
The deployment was just about to succeed, but then aborted with the
message "Aborted!" when the confirmation prompt appeared.

Per https://github.com/aws/aws-sam-cli/issues/2180, I _could_ remove
'confirm_changeset = true' from samconfig.toml, which is used by local
'sam deploy' runs. For now, I want to keep that file as it is, and just
use the explicit '--no-confirm-changeset' flag in the pipeline.yaml
file.