Add CGO_ENABLED=0 to build-EListMan make target
The CI/CD pipeline is now failing in the integration-test phase for two
reasons. One is that the 'curl' commands in bin/smoke-test.sh are
malformed. I'll address that in the next commit.

In this commit, I address the error from the CloudWatch logs for the
elistman-dev deployment:

  /var/task/main: /lib64/libc.so.6: version `GLIBC_2.32' not found
  (required by /var/task/main)

This seems to be why the Lambda docs suggest setting CGO_ENABLED=0 in
the first place, though it's never been a problem until now:

- https://docs.aws.amazon.com/lambda/latest/dg/golang-package.html