Add SNS topic, SES config set for delivery events
This is in preparation for having the Lambda function handle SES
delivery event notifications for bounces, complaints, etc.

I recently realized that it's up to the EListMan application to manage
this, since we're not using AWS SES list management. Once I successfully
deploy the infrastructure, I'll turn to handling the events. Hopefully
it won't take as much time and effort as handling HTTP API and mailto
events.