Make SNS topic policy Action a list, update ID
The previous attempt to deploy failed:

- https://github.com/mbland/elistman/actions/runs/4661623744/jobs/8251134356

  UPDATE_FAILED
  AWS::SES::ConfigurationSetEventDestination
  SendingConfigurationSetDestination
  Resource handler returned message: "Could not publish message to SNS
  topic <arn:aws:sns:us-east-1:***:elistman-dev>. (Service: SesV2,
  Status Code: 400, Request ID: ...)" (RequestToken: ...,
  HandlerErrorCode: InvalidRequest)

I think it might be because "Action" should've been a list of string,
not just a string. We'll see.

Also decided to give the PolicyDocument a unique ID, to possibly avoid a
similar problem to the conflict noted in commit
d85327c08a90465e150442b52208d353a8ea9744.