Restore config set destination, update SNS policy
After deploying the previous commit, the SNS topic policy looked as
expected. However, I finally stumbled upon the page showing the policy
for an event destination, which is slightly different than the policy
for verified identity notifications.

In other words, I was previously following the instructions here:
- https://docs.aws.amazon.com/ses/latest/dg/configure-sns-notifications.html

when I needed to follow the instructions here:
- https://docs.aws.amazon.com/ses/latest/dg/event-publishing-add-event-destination-sns.html

As noted in the comment I added to template.html, the difference is
subtle. It's possible to have one topic for multiple deployments using
the verified identity notifications. For this application, however, we
set up different topics and destinations for each deployment. This keeps
each deployment's notifications encapsulated from one another.