Add SES+SNS event notification handlers
Most of these events will indicate that the recipient should be removed
from the subscriber list. The one exception is the "no-spam" complaint,
which indicates that the recipient should be restored to the list.

Note that the TestNewSesEventHandler test cases essentially test the
JSON unmarshaling of the handler/ses_event.go types.