Add DynamoDb.{Create,Delete}Table, randomString()
db/database_test.go is now successfully creating and removing a test
table with a randomly generated name.

Making CreateTable and DeleteTable members of DynamoDb (instead of
separate test helpers) provides a couple of benefits:

- The same DynamoDb class can be used to manage local and remote
  DynamoDB test tables.

- I can use it in the future to provide a create-table CLI to replace
  bin/create-subscribers-table.sh, or to create the table automatically
  on startup if necessary.

I'm also thinking of matching the table name to the SAM/CloudFormation
stack name, with an "elistman-" prefix. Consistent use of this
convention would reduce the amount of explicit administrative setup and
coordination required. It would also make it easy to tell which resources
belong to which EListMan instance via either the AWS CLI or web console.

This change also changed the type of the first parameter of
NewDynamoDb() from aws.Config to *aws.Config. dynamodb.NewFromConfig()
may have a reason to take the config parameter by copy. But for our own
client code, passing a pointer through up to that point seems to make
sense.