Implement DynamoDb, full test suite, enum parser
It needs refactoring, but this is essentially a complete implementation
of both DynamoDb and it test suite. The tests still only use a local
DynamoDB via Docker for now. However, if that implementation upholds its
contract, it should be easy to run the same tests against an actual
remote DynamoDB table.

As part of the implementation, I needed to add a parser to parse stored
SubscriberStatus string values back to a SubscriberStatus. This parser
essentially reverses the SubscriberStatus.String() implemented by the
stringer generator. Maybe it's something I could add back upstream some
day.