Extract dynamodb{,_contract_test}.go
These new files make the organization of the code and intent of the
tests more explicit.

It occurred to me that we may actually want to collect code coveage from
contract tests, at least as I prefer to define the term. Specifically,
if the contract test is focused on exercising an "internal API" wrapping
an external dependency, having coverage for that wrapper can be helpful.
Hence, I've introduced the 'contract_tests' build tag, which I may put
to use soon.