Collect coverage from contract_tests
This updates both 'make coverage' and the GitHub run-tests pipeline. As
long as the contract test exercises a thin internal API layer, it make
sense to collect coverage for that layer. (I do need to add some extra
coverage for the db package still; the overall coverage will probably
drop slightly in Coveralls for now.)