First broken attempt at email.AddressValidatorImpl
Though I'm abandoning this approach, I want to keep this commit in the
commit history for reference.

My plan was to implement the algorithm described here for checking SMTP
servers for valid email addresses before sending:

- https://mailtrap.io/blog/verify-email-address-without-sending/

The code itself works as expected, but it seems most major modern MX
servers don't allow anyone to dial in and ping mailboxes anymore.

In the process of trying to solve this problem, I've read about and
discovered all kinds of email security protocols and utilities,
including:

- https://github.com/zone-eu/mx-connect
- https://github.com/postalsys/mailauth
- https://github.com/jetmore/swaks
- https://toolbox.googleapps.com/apps/checkmx/

I tried to gain insights from reading the Postfix source from
https://www.postfix.org. I learned about Strict Transport Security,
which I thought might be part of the problem:

- https://www.mailhardener.com/kb/mta-sts
- RFC 8461: SMTP MTA Strict Transport Security (MTA-STS):
  https://www.rfc-editor.org/rfc/rfc8461
- About MTA-STS and TLS reporting:
  https://support.google.com/a/answer/9261504
- What is MTA-STS and Why Do You Need It?:
  https://powerdmarc.com/what-is-mta-sts-and-why-do-you-need-it/

I also learned about Simple Authentication and Security Layer (SASL) for
SMTP, or SASL, which definitely isn't part of the problem:

- RFC 4422: Simple Authentication and Security Layer (SASL):
  https://www.rfc-editor.org/rfc/rfc4422
- RFC 4954: SMTP Service Extension for Authentication:
  https://datatracker.ietf.org/doc/html/rfc4954
- https://www.postfix.org/SASL_README.html
- https://www.cyrusimap.org/sasl/
- https://github.com/cyrusimap/cyrus-sasl
- RFC 4752: The Kerberos V5 ("GSSAPI") Simple Authentication and
  Security Layer (SASL) Mechanism:
  https://www.rfc-editor.org/rfc/rfc4752

But in the end, none of these tools could connect to the Gmail servers
or the mailroute.net servers used by acm.org. It seems most (major,
legit) MX servers greatly restrict incoming SMTP traffic—and justifiably
so:

- https://serverfault.com/questions/419407/fighting-spam-what-can-i-do-as-an-email-administrator-domain-owner-or-user