Test, refactor AddressValidator; add suppression
AddressValidator is completely tested except for the final branch of
ValidateAddress when checkMailHosts fails. However, I want the
changes to test that branch to appear by themselves in the next commit,
which will also add contract test support.

This change introduces the Suppressor.Suppress method, since I realized
that we don't want to repeat DNS validation for known bad addresses.

The algorithm and its outcomes are otherwise largely unchanged, but
heavily refactored for improved readability. In particular, the generic
lookup() function eliminates boilerplate around LookupMX, LookupAddr,
and LookupHost. Also, several of the method names are much clearer, as
are several error messages. Hopefully the logic of each method is much
easier to follow, as some logic has been moved from some methods to
others.

It does extra time and effort to improve the quality of some working
code sometimes. But the results are so worth it.