Remove ReceiptRuleSet from template, add to README
Today I learned that only one ReceiptRuleSet can be active for an
account, so this is something that the EListMan SAM template shouldn't
manage. The README now describes how to set up an active ReceiptRuleSet,
SNS, and MX records. The bin/sam-with-env.sh script now requires the
RECEIPT_RULE_SET_NAME variable.