Replace SESCrudPolicy with policy Statement
This is something I borrowed from my recent side project,
https://github.com/mbland/ses-forwarder. Per commit
992e55323ca42c2f1eb7a0c97d164a27a96b5396 in that project, SESCrudPolicy
isn't permissive enough to allow access to the configuration set
generated by this template.

Even so, I'm wondering if this is still too permissive. I wonder if it's
better to set specific, separate policies for both the SES verified
identity and the generated configuration set.