PoC: SesMailer, MessageBuilder, Writer, first mail
This is an initial proof of concept for generating MIME
multipart/alternative emails with a text part and an email part. The
TestBuildMessage test fails, but the failure message shows that the
message looks as expected.

The problem is—and this may be premature optimization—it seems wasteful
to re-encode the entire multipart message for each individual
subscriber.

I actually had an idea overnight for solving this problem that will be
in the next commit.