Add Message type, use template to avoid reencoding
This is another proof of concept with a failing test whose failure
message shows it working, just like the previous commit.

The idea I had to avoid reencoding the entire message for each
subscriber was to introduce the Message type. This type contains
separate body and footer fields for both plaintext and HTML. We only
encode the bodies once, then reencode only the footers, which should
contain the UnsubscribeUrlTemplate, for each subscriber.

This commit also shows some refactoring to push behavior into the
Subscriber and Message objects. The next commit will polish the code and
implement the tests properly.