Refactor convertToCrlf
It was bugging me that the returned []byte might hold onto extra,
unnecessary memory. The counting of newlines also started to bother me.
So I refactored it to allocate a buffer twice the size of the original,
and then returned a buffer trimmed to the necessary size.

Also defines newline and carriageReturn byte constants.

This is partly avoidance. I need to finish testing the MessageTemplate.