Test emitPart (MessageTemplate helper)
This required some slight reconfiguration of some of the existing test
data. This commit repurposes much of the former textContent as
encodedTextFooter and replaces textContent with textOnlyContent.

instantiatedTextFooter was extracted from encodedTextFooter.
instantiatedTextFooter is one of the inputs to the new TestEmitPart.
encodedTextFooter is part of the new textPart, used as the expected
result in TestEmitPart/Succeeds.

The upcoming TestMultipart will use textPart, and the upcoming
TestEmitMessage will use both textOnlyContent and textPart.

fixup