WIP: First TestEmitMultipart case, more helpers
This is another Exploration-phase-headed-towards-Settlement-and-Vision
commit. It's not finished and polished, but contains seeds of important
work that I plan to polish and complete shortly.

I've added more test variables for the HTML output, and added a couple
of helpers for parsing the resulting message parts. Specifically:

- parseContentAndMediaType
- parseAndCheckDecoded
- checkNextPart

There's some degree of duplication between the first two helpers and the
third that I hope to eliminate very shortly.