Update db.Subscriber.Verified to Status, timestamp
This is in preparation for the upcoming Global Secondary Index scheme,
whereby one GSI will contain "pending" records and the other "verified"
records. "pending" will be the partition key for one, and "verified"
will be the partition key for the other. This way scanning all the
"verified" records won't incur the cost of scanning "pending" records.

There's also no need to store or read a "Verified" boolean property any
longer, as each record can contain only a "pending" or "verified"
attribute.

The "pending" timestamp will also serve as a DynamoDB Time To Live
value:

- https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/TTL.html

This way, DynamoDB will automatically expire and remove any stale
subscriber records stuck in the "pending" state.

This commit contains a few constant and test variable renames and
relocations. It also contains the new Subscriber.String method and the
toDynamoDbTimestamp helper, the latter physically colocated with
dbParser.GetTime().