Add indexes, DynamoDb.GetSubscribersInState
Adds two Global Secondary Indexes (GSIs), one containing "pending"
records and the other "verified" records, and the GetSubscribersInState
method for scanning either GSI.

- https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/GSI.html

Each Subscriber can only contain either a "pending" timestamp or a
"verified" timestamp, not both. The "pending" timestamp is the partition
key for one GSI, and the "verified" timestamp is the partition key for
the other. As a result, each record can only appear in one GSI or the
other. This way scanning all the "verified" records in order to send an
email to the list won't incur the cost of scanning the "pending"
records.

This is the same principle as "sparse indexes," though in this case, we
expect most of the table entries to be "verified." Still, it seems worth
avoiding the extra scan and filtering overhead.

- https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/bp-query-scan.html
- https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/bp-indexes-general-sparse-indexes.html

"pending" timestamp as TTL
--------------------------

As mentioned in the previous commit, the "pending" timestamp will also
serve as a DynamoDB Time To Live value:

- https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/TTL.html

Support for this is forthcoming in a future commit.

Waiting for GSIs in dynamodb_contract_test
------------------------------------------

Adding the two GSIs to the test table slows down the
dynamodb_contract_test considerably when running against the actual
DynamoDB (via the -awsdb flag). Consequently, the default value for
-dbwaitattempts is now 12, instead of 3, allowing for a full minute to
wait for table creation by default. The test now completes on the order
of 20 to 30-some seconds.

A couple of the new GetSubscribersInState tests originally failed once
because the table or GSIs didn't update fast enough. For this reason,
waitIfTestingAgainstAws introduces a three second delay in a couple of
places when -awsdb is set.

Convention for sharing test objects (and related build tags issue)
------------------------------------------------------------------

This also moves some test objects to db/db_test.go, which contains no
tests itself. I think I'm going to use this convention moving forward,
i.e., putting all shared test objects in $PACKAGE/$PACKAGE_test.go. I'll
rename email/utils_test.go to email/email_test.go accordingly later.

A problem with these common files is that, without build tags, VS Code's
go-staticcheck will report unused objects. The command line 'go run
.../staticcheck -tags=all_tests ./...' (from 'make static-checks') runs
fine in either case. It seems silly to add build tags to these files
when, without them, they should be accessible to all build
configurations already. However, at least its an easy workaround.