Add DynamoDb.UpdateTimeToLive
As mentioned in the previous commit, the "pending" timestamp now serves
as a DynamoDB Time To Live value:

- https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/TTL.html

The upcoming CLI will use the pattern from setupDynamoDb() in
dynamodb_contract_test to create and wait for the table. Then it will
call UpdateTimeToLive on it, so that "pending" subscribers that don't
become "verified" by a deadline are automatically expired and removed.

The ProdAgent will have to set the "pending" timestamp for a new
subscriber at some point in the future, likely 24 hours.