Implement email.SesMailer email.Suppressor methods
SesMailer.IsSuppressed and SesMailer.Suppress are fully implemented and
tested, as is the new SesMailer.Unsuppress. Added
mailer_contract_test.go to exercise these against the actual AWS SESV2
API.

mailer_contract_test actually found a bug that the unit test didn't
catch, when unexpected "unexpected error" log messages appeared. This
led me to discover the following API documentation, which inspired the
correct errors.As method of checking Go client errors:

- https://aws.github.io/aws-sdk-go-v2/docs/handling-errors/#api-error-responses

as opposed to the previous type assertion implementation:

  if _, ok := err.(*typesV2.NotFoundException); !ok {

The errors.As update was compatible with the first version of the
TestIsSuppressed/ReturnsFalse/IfNotSuppressed unit test. The test has
since been updated to wrap a NotFoundException, instead of returning it
directly, to ensure that the implementation uses errors.As properly.

While this normally may be an implementation detail we'd rather not
expose, this is how the aws-sdk-go-v2 API is supposed to work. Hence,
it's important to verify that we're using it properly—as the contract
test showed that we weren't before.

This commit also contains a few additional minor changes:

- Extracted testutils.LoadDefaultAwsConfig
- Added testutils.Logs.Reset (though I didn't end up using it yet)
- Added testutils.RandomEmail

fixup