Add TestDynamoDbClient, move GetSubscribers tests
Per the TestDynamoDbClient comment, most of its methods are
unimplemented, because dynamodb_contract_test.go tests most of them.

The exception to this is Scan(), which is the reason why the
DynamoDbClient interface exists. Testing all the cases of the code that
relies on Scan() is annoying, difficult, and/or nearly impossible
without using this test double.

Case in point: Compare the tests for GetSubscribersInState moved from
dynamodb_contract_test.go to dynamodb_test.go in this commit. Though
this required adding logic for TestDynamoDbClient.Scan, the tests are
faster and more reliable. As we'll see with the upcoming
ProcessSubscribersInState, testing all its cases will prove _much_
easier and more reliable with the TestDynamoDbClient approach.