Apply ops.AwsError throughout db, email packages
The new tests ensure that everywhere a server error could occur, that
ops.AwsError wraps it with ops.ErrExternal.

After making this change, I'm getting a mind to update the AwsError API
to avoid having to wrap the error every time. I think there's a need for
a new IsErrExternal test helper as well.