Update ops.AwsError to take a string prefix
As mentioned in the previous commit, this eliminates the need to wrap an
AWS error to add a prefix to it before calling ops.AwsError.

This means the error returned from ops.AwsError will only wrap
ops.ErrExternal, not any original AWS errors. This is good because it
narrows the scope of the API; i.e., if we don't returned wrapped AWS
errors, clients can't depend on them.

Plus, the new API is more convenient for building error messages,
eliminating the extra fmt.{Errorf,Sprintf} in most cases.