Use &Subscriber in test, not newTestSubscriber
This was working in VS Code because go.buildTags is set to all_tests in
.vscode/settings.json. Every test file has this build tag defined, so
newTestSubsciber from dynamodb_contract_test.go was available to
dynamodb_test.go.

When trying to run `make small-tests`, however, it failed with:

  go test -tags=small_tests ./...
  # github.com/mbland/elistman/db [github.com/mbland/elistman/db.test]
  db/dynamodb_test.go:207:23: undefined: newTestSubscriber
  FAIL    github.com/mbland/elistman/db [build failed]

TestDynamodDbMethodsReturnExternalErrorsAsAppropriate doesn't need a
fully initialized Subscriber object. Replacing newTestSubscriber with an
empty &Subscriber fixes the problem.