Begin implementing, testing ProdAgent.Subscribe
ProdAgent.Subscribe now:

- validates the email address
- tries to get an existing subscriber record and creates a new one if no
  such record exists
- sets the Uid and Timestamp on the subscriber record
- writes the subscriber record to the database

The one thing it doesn't do yet is send the verification email, which
will come in a subsequent commit.

This commit introduces the testdoubles package with AddressValidator and
Database implementations used to implement the tests in
agent/agent_test.go.

A notable insight from the ProdAgent.Subscribe comments:

  ValidateAddress returns an error describing why an address failed
  validation. However, this function should only return errors that are
  logical or infrastructure failures. Hence we log the result, but don't
  return the error.

  Perhaps ValidateAddress needs a better API, returning a failure reason
  separately from a logical/infrastructure error.