Rename *Agent.UnsubscribeBaseUrl to ApiBaseUrl
main.go was already initializing it correctly as an ApiBaseUrl. The
UnsubscribeBaseUrl name came from when I needed to pass it down to
email.Subscriber.SetUnsubscribeInfo.

Speaking of which, that function needs updating. But first, I need to
move the API path prefixes out of handler/parser.go to another module to
avoid a circular dependecy.