Make Unsuppress pass for nonexistent user
It occurred to me just before implementing agent.ProdAgent.Restore that
SesSuppressor.Unsuppress could return successfully when trying to
"unsuppress" addresses not on the suppression list. This fits with the
"Easier to Ask For Permission" pattern from:

- https://docs.python.org/3/glossary.html#term-EAFP

I've heard this is good concurrency and security practice, too, as it
avoids races between the "checking" and "executing" steps, a.k.a. "Look
Before You Leap":

- https://docs.python.org/3/glossary.html#term-LBYL