Add agent.ProdAgent.{Remove,Restore}
This completes the agent.ProdAgent implementation—and that of all the
core functionality of EListMan!

Except for one more thing...a command line interface for sending an email to
the list. I'll also have to disable some smoke tests before switching
lambda/main.go to use the ProdAgent instead of the DecoyAgent. Then it's
essentially ready to ship!

This commit adds an email.Suppressor to the ProdAgent, et. al., and the new
testdoubles.Suppressor. It came together pretty quickly and easily, though.

I also learned that DynamoDB doesn't raise an error when trying to delete a
nonexisting item. That makes the API much easier to use. Confirmed by adding
an extra check to dynamodb_contract_test.