Return AlreadySubscribed for pending subscribers
It occurred to me that someone could try to use the Subscribe function
to repeatedly subscribe themselves or someone else while the user is
pending. So now the policy is that you only get one verification email,
until the record expires (thanks to DynamoDB's Time-To-Live feature) or
the adminstrator intervenes.

In the process, removed agent.ProdAgent.getOrCreateSubscriber and its
test. Also added testdoubles.Mailer.AssertNoMessageSent.