Add processDnsError for email.AddressValidator
I suspected that there must be a better way of detecting whether a
failed net.Resolver lookup was due to an external error. After studying
the source for the net library, I realized the following, per the
processDnsError comments:

- All errors are of type net.DNSError.
- If a host is found, it will be returned, even if there's a non-nil
  error accompanying it in some cases. In that case, do not call this
  function.
- If there were no problems with the network or DNS response, but the
  host was not found, no hosts are returned. The error will be a
  net.DNSError with IsNotFound == true.
- If there were network or DNS issues, no hosts are returned, and the
  error will be a net.DNSError with IsNotFound == false.

Hence, processDnsError returns nil if err is a DNSError and IsNotFound
is true. Otherwise it presumes the error is a network or other external
failure and wraps it with ops.ErrExternal.

This function will be applied to the email.AddressValidator
implementation in a subsequent commit.