Add "\n" to text and HTML email bodies if needed
This adds a trailing newline to the textBody or htmlBody fields of an
email.Message if it's missing before convertToCrlf in
NewMessageTemplate.

I generated a test email using an early prototype of the CLI and noticed
lines over 76 characters wide. Since the email.MessageTemplate contents
are quoted-printable encoded, this shouldn't've happened.

I realized the real problem was that my test input didn't end its text
and HTML bodies with a newline. This change fixes the issue.