Add initial Cobra CLI with 'preview' command
Cobra is a popular framework for writing CLI programs in Go:

- https://github.com/spf13/cobra
- https://github.com/spf13/cobra-cli/blob/main/README.md
- https://github.com/spf13/cobra/blob/main/user_guide.md

Right now I'm just running the command via:

$ go run ./main.go
$ go run ./main.go preview < test.json

Seeing a preview of the message using an earlier prototype of the
preview command led to the fix in commit
8cd3a37b42cafcfe0fe2e2374aeb7cc8c2587da3.

I'm also thinking the 'elistman send' command will invoke the remote
Lambda to send a message, versus sending it locally.

This is due to the fact that several bits of configuration are needed to
send a message, and the deployed Lambda already has them. Adding an
event to the Lambda handler that's provided by a direct CLI invocation
seems better, less complicated, annoying, and error prone.