Add MessageTemplate.GenerateMessage, update tests
After working on building a message for ProdAgent.Send, I realized that
the common case for generating MessageTemplate messages was to write to
a bytes.Buffer. Writing to this buffer cannot return an error; it can
only panic if and when it runs out of memory.

Adding GenerateMessage provides convenience for this common case and
made several of the tests shorter and easier to understand.

The fact that the EmitMessage call tree exhaustively checks errors isn't
for naught; it's still used directly by the 'preview' command to write
to os.Stdout.

Also updated newTestRecipient() and several Recipient variable names
while I was at it.