Extract EmitPreviewMessageFromJson
At the moment, the 'preview' command is essentially a one-liner.

In the process of extracting and testing this, I realized that
MessageTemplate.EmitMessage wasn't passing through a wrapped error. That
might've been deliberate, but I'm more inclined to wrapping by default
than not these days.

I'm also thinking of extracting a lot of the testutils/email.go logic
into a proper message parsing library. Then the 'elistman preview'
command can emit previews of the text and HTML parts of the message
after {{UnsubscribeUrl}} substitution and encoding/decoding.